+++ packages/medusa-plugin-restock-notification/src/services/__tests__/restock-notification.js
@@ -147,17 +147,9 @@ describe("RestockNotificationService", () => {
         eventBusService: EventBusService,
       })
 
-      restockNotiService.restockExecute_ = jest.fn()
-
-      jest.clearAllMocks()
-      jest.useFakeTimers()
-
+      restockNotiService.restockExecute = jest.fn()
       restockNotiService.triggerRestock("variant_test")
-
-      jest.runAllTimers()
-
-      expect(setTimeout).toHaveBeenCalledTimes(1)
-      expect(setTimeout).toHaveBeenCalledWith(expect.any(Function), 0)
+      expect(restockNotiService.restockExecute).toHaveBeenCalledTimes(1)
     })
 
     it("trigger delay 10", async () => {
@@ -171,21 +163,20 @@ describe("RestockNotificationService", () => {
         { trigger_delay: 10 }
       )
 
-      restockNotiService.restockExecute_ = jest.fn()
-
-      jest.clearAllMocks()
-      jest.useFakeTimers()
+      restockNotiService.restockExecute = jest.fn()
 
       restockNotiService.triggerRestock("variant_test")
 
-      jest.runAllTimers()
-
-      expect(setTimeout).toHaveBeenCalledTimes(1)
-      expect(setTimeout).toHaveBeenCalledWith(expect.any(Function), 10)
+      expect(EventBusService.emit).toHaveBeenCalledTimes(1)
+      expect(EventBusService.emit).toHaveBeenCalledWith(
+        "restock-notification.execute",
+        { variant_id: "variant_test" },
+        { delay: 10 }
+      )
     })
   })
 
-  describe("restockExecute_", () => {
+  describe("restockExecute", () => {
     const restockNotiService = new RestockNotificationService({
       manager: MockManager,
       productVariantService: ProductVariantService,
@@ -196,20 +187,20 @@ describe("RestockNotificationService", () => {
     it("non-existing noti does nothing", async () => {
       jest.clearAllMocks()
 
-      await expect(restockNotiService.restockExecute_("variant_test")).resolves
+      await expect(restockNotiService.restockExecute("variant_test")).resolves
     })
 
     it("existing noti but out of stock does nothing", async () => {
       jest.clearAllMocks()
 
-      await expect(restockNotiService.restockExecute_("variant_outofstock"))
+      await expect(restockNotiService.restockExecute("variant_outofstock"))
         .resolves
     })
 
     it("existing noti emits and deletes", async () => {
       jest.clearAllMocks()
 
-      await restockNotiService.restockExecute_("variant_1234")
+      await restockNotiService.restockExecute("variant_1234")
 
       expect(EventBusService.emit).toHaveBeenCalledTimes(1)
       expect(EventBusService.emit).toHaveBeenCalledWith(
@@ -238,7 +229,7 @@ describe("RestockNotificationService", () => {
         { inventory_required: 5 }
       )
 
-      await service.restockExecute_("variant_1234")
+      await service.restockExecute("variant_1234")
 
       expect(EventBusService.emit).toHaveBeenCalledTimes(1)
       expect(EventBusService.emit).toHaveBeenCalledWith(
@@ -265,7 +256,7 @@ describe("RestockNotificationService", () => {
         { inventory_required: 5 }
       )
 
-      await service.restockExecute_("variant_low_inventory")
+      await service.restockExecute("variant_low_inventory")
 
       expect(EventBusService.emit).toHaveBeenCalledTimes(0)
       expect(RestockNotificationModel.delete).toHaveBeenCalledTimes(0)