+++ packages/medusa/src/api/routes/admin/shipping-options/__tests__/create-shipping-option.js
@@ -43,8 +43,9 @@ describe("POST /admin/shipping-options", () => {
         name: "Test option",
         region_id: "testregion",
         provider_id: "test_provider",
+        metadata: undefined,
         data: { id: "test" },
-        profile_id: expect.stringMatching(/.*/),
+        profile_id: expect.any(String),
         price_type: "flat_rate",
         amount: 100,
         requirements: [
@@ -86,7 +87,9 @@ describe("POST /admin/shipping-options", () => {
     })
 
     it("returns error", () => {
-      expect(subject.body.message[0].message).toEqual(`"name" is required`)
+      expect(subject.body.message).toEqual(
+        expect.stringContaining(`name must be a string`)
+      )
     })
   })
 })