+++ packages/medusa/src/api/routes/admin/shipping-options/get-shipping-option.ts
@@ -3,6 +3,7 @@
  * operationId: "GetShippingOptionsOption"
  * summary: "Retrieve a Shipping Option"
  * description: "Retrieves a Shipping Option."
+ * x-authenticated: true
  * parameters:
  *   - (path) id=* {string} The id of the Shipping Option.
  * tags: