+++ packages/medusa/src/models/region.ts
@@ -35,7 +35,7 @@ export class Region {
   @JoinColumn({ name: "currency_code", referencedColumnName: "code" })
   currency: Currency
 
-  @Column({ type: "decimal" })
+  @Column({ type: "real" })
   tax_rate: number
 
   @Column({ nullable: true })