+++ packages/medusa/src/services/__mocks__/invite.js
@@ -1,0 +1,23 @@
+export const InviteServiceMock = {
+  list: jest.fn().mockImplementation((selector, config) => {
+    return Promise.resolve({})
+  }),
+
+  create: jest.fn().mockImplementation(data => {
+    return Promise.resolve({})
+  }),
+
+  accept: jest.fn().mockImplementation((token, user_id) => {
+    return Promise.resolve({})
+  }),
+
+  resend: jest.fn().mockImplementation((id, inviter) => {
+    return Promise.resolve({})
+  }),
+}
+
+const mock = jest.fn().mockImplementation(() => {
+  return InviteServiceMock
+})
+
+export default mock